requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   langchain-community
    #   voyageai
aiolimiter==1.2.1
    # via voyageai
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
apscheduler==3.11.1
    # via flask-apscheduler
attrs==25.4.0
    # via aiohttp
bcrypt==5.0.0
    # via -r requirements.in
beautifulsoup4==4.14.2
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   flask
    #   geocoder
    #   typer-slim
cloudscraper==1.2.71
    # via -r requirements.in
dataclasses-json==0.6.7
    # via langchain-community
decorator==5.2.1
    # via ratelim
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via pymongo
faker==37.12.0
    # via -r requirements.in
feedparser==6.0.12
    # via -r requirements.in
filelock==3.20.0
    # via huggingface-hub
flask==3.1.2
    # via
    #   flask-apscheduler
    #   flask-pymongo
flask-apscheduler==1.13.1
    # via -r requirements.in
flask-pymongo==3.0.1
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via huggingface-hub
future==1.0.0
    # via geocoder
geocoder==1.38.1
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   huggingface-hub
    #   langsmith
    #   openai
httpx-sse==0.4.3
    # via langchain-community
huggingface-hub==1.0.1
    # via tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.11.1
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via
    #   langchain-community
    #   langchain-mongodb
langchain-community==0.3.31
    # via -r requirements.in
langchain-core==0.3.79
    # via
    #   langchain
    #   langchain-community
    #   langchain-mongodb
    #   langchain-openai
    #   langchain-text-splitters
    #   langchain-voyageai
langchain-mongodb==0.7.2
    # via -r requirements.in
langchain-openai==0.3.35
    # via -r requirements.in
langchain-text-splitters==0.3.11
    # via
    #   langchain
    #   langchain-mongodb
    #   voyageai
langchain-voyageai==0.1.7
    # via -r requirements.in
langsmith==0.4.38
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lark==1.3.1
    # via langchain-mongodb
markdown==3.9
    # via -r requirements.in
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
marshmallow==3.26.1
    # via dataclasses-json
mistune==3.1.4
    # via -r requirements.in
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.3
    # via
    #   langchain-community
    #   langchain-mongodb
    #   voyageai
openai==2.6.1
    # via langchain-openai
orjson==3.11.4
    # via langsmith
packaging==25.0
    # via
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
pillow==12.0.0
    # via
    #   -r requirements.in
    #   voyageai
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycountry==24.6.1
    # via -r requirements.in
pydantic==2.12.3
    # via
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langchain-voyageai
    #   langsmith
    #   openai
    #   pydantic-settings
    #   voyageai
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via langchain-community
pydub==0.25.1
    # via -r requirements.in
pymongo==4.15.3
    # via
    #   flask-pymongo
    #   langchain-mongodb
pyparsing==3.2.5
    # via cloudscraper
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   flask-apscheduler
python-dotenv==1.2.1
    # via pydantic-settings
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
qrcode==8.2
    # via -r requirements.in
ratelim==0.1.6
    # via geocoder
regex==2025.10.23
    # via tiktoken
requests==2.32.5
    # via
    #   cloudscraper
    #   geocoder
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
    #   voyageai
requests-toolbelt==1.0.0
    # via
    #   cloudscraper
    #   langsmith
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via huggingface-hub
six==1.17.0
    # via
    #   geocoder
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via
    #   langchain
    #   langchain-community
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
    #   voyageai
tiktoken==0.12.0
    # via langchain-openai
tokenizers==0.22.1
    # via voyageai
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   beautifulsoup4
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer-slim
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via faker
tzlocal==5.3.1
    # via apscheduler
urllib3==2.5.0
    # via requests
uwsgi==2.0.30
    # via -r requirements.in
voyageai==0.3.5
    # via langchain-voyageai
websocket-client==1.9.0
    # via -r requirements.in
werkzeug==3.1.3
    # via flask
yarl==1.22.0
    # via aiohttp
zstandard==0.24.0
    # via langsmith
